==> code/01_identify_sibs.Rmd <==
---
title: "Identify Siblings"
author: Casey Breen
---

Summary: This notebook contains code to identify brothers in census records by linking the 1920 census to the 1940 census. 

```{r}
## Load helper functions and library all packages used for analysis
library(here)
source(here("code/helpers.R"))
```

## Read in data 

```{r}
## read in ABE crosswalk 
xwalk_1920_1940 <- fread("/90days/casey/1920-1940/1920-1940/data/crosswalk_1920_1940.csv")

## read in 1920 census 
census_1920 <- fread("/home/ipums/casey-ipums/IPUMS2019/1920/TSV/P.tsv", select = c("AGE", "SEX", "HISTID", "SERIALP", "RELATE")) %>% 
  janitor::clean_names()

## read in 1940 census linked to 1920 census 
censoc_dmf <- fread("/censoc/data/censoc_linked_with_census/1940_v2.1/censoc_dmf_v2_1_linked_with_census.csv")
```

## Identify Siblings

```{r}
## identify men age 5-16 in the 1940 census 
siblings_1920 <- census_1920 %>% 
  filter(age %in% c(3:17)) %>% 
  filter(sex == 1) %>% 
  rename_all(paste0, "_1920")

## calculate birth order 
siblings_1920 <- siblings_1920 %>%
  arrange(serialp_1920, -age_1920) %>%
  group_by(serialp_1920) %>%
  mutate(birth_order = row_number())  

## match 1920 and 1940 census 
siblings_matched <- siblings_1920 %>% 
  inner_join(xwalk_1920_1940, by = "histid_1920") %>% 
  filter(link_abe_exact_standard == 1) %>% 
  rename(link_abe_exact_conservative_1920_1940 = link_abe_exact_conservative)

## identify siblings (already restricted to link_abe_exact_standard == 1)
censoc_dmf_linked <- censoc_dmf %>% 
  rename(serialp_1940 = serialp) %>% 
  rename(link_abe_exact_conservative_1940_dmf = link_abe_exact_conservative) %>% 
  inner_join(siblings_matched, by = c("histid" = "histid_1940")) 

## identify linked siblings  
censoc_dmf_linked_sibs_final <- censoc_dmf_linked %>% 
  filter(byear %in% 1905:1915) %>% 
  filter(age_1920 %in% 4:15) %>% 
  filter(age %in% 24:35) %>% 
  filter(bpl < 15000) %>% 
  filter(relate == 101) %>% 
  filter(ownershp != 0) %>% 
  group_by(serialp_1920) %>% 
  mutate(n = n()) %>% 
  filter(n > 1) %>% 
  ungroup()

## recode education 
censoc_dmf_linked_sibs_final <- censoc_dmf_linked_sibs_final %>% 
  censocdev::recode_education(educ_var = "educ") %>% 
  mutate(ownership = relevel(as.factor(ownershp), ref = "20")) %>% 
  mutate(racec = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black",
    TRUE ~ "Other"
  )) %>% 
  mutate(educ_yrs = case_when(
    is.na(educ_yrs) ~ median(educ_yrs, na.rm = T),
    TRUE ~ educ_yrs))

## eldest sib and first born 
censoc_dmf_linked_sibs_final <- censoc_dmf_linked_sibs_final %>% 
  group_by(serialp_1920) %>% 
  mutate(first_born = case_when(
    birth_order == 1 ~ 1, 
    TRUE ~ 0)) %>% 
  mutate(eldest_brother = case_when(
    birth_order == min(birth_order) ~ 1,
    TRUE ~ 0
  ))
```


```{r}
## write out linked census file
write_rds(censoc_dmf_linked_sibs_final, here("data/dmf_sibs.rds"))
```
