==> code/02_home_ownership.Rmd <==
---
title: "Home Ownership Analysis"
author: Casey Breen
---


Summary: code to look at relationship between home ownership and longevity. 

```{r}
## library packages
library(data.table)
library(tidyverse)
library(ipumsr)
library(socviz)
library(broom)
library(cowplot)
library(fixest)
```


```{r}
## read in dmf 
dmf <- fread("/data/josh/CenSoc/censoc_data/censoc_linked_to_census/censoc_dmf_v2_linked.csv")

## read in IPUMS DDI for IPUMS extract
## we already have a ddi on the server for all vars! but can also make a new extract from IPUMS with BPLD var  
ddi_extract <- read_ipums_ddi("/data/josh/CenSoc/censoc_data/ipums_1940_extract/fullcount.ddi.xml")

## rename our bpl var to match the IPUMS var BPLD

## assign metadata using ipumsr package
dmf <- ipums_collect(data = dmf, ddi = ddi_extract, var_attrs = c("val_labels", "var_label", "var_desc"))

## create new string variable bpl_string 
dmf <- dmf %>% 
  mutate(bpl_string = as_factor(BPLD))
```




```{r}
dmf_analysis <- dmf %>% 
  filter(byear %in% 1905:1915) %>% 
  janitor::clean_names() %>% 
  mutate(ownership = relevel(as.factor(ownershp), ref = "2")) %>% 
  filter(ownership != 0) %>% 
  censocdev::recode_education(educ_var = educd)

dmf_analysis <- dmf_analysis %>% 
  filter(link_abe_exact_conservative == 1)

test <- feols(death_age ~ ownership + educ_yrs | byear + urban + occ + race + marst + statefip,  data = dmf_analysis %>% 
                filter(relate == 1))

test1 <- feols(death_age ~ ownership | byear,  data = dmf_analysis %>% 
  filter(relate == 1))

test2 <- feols(death_age ~ ownership | byear,  data = dmf_analysis_sibs)
```


```{r}
## read in mortality data sibs 
mortality_data <- fread("../data/censoc_sibs.csv") %>% 
  janitor::clean_names()

dmf_analysis_sibs <- mortality_data %>% 
  mutate(ownership = relevel(as.factor(ownershp), ref = "2")) %>% 
  filter(ownership != 0) %>% 
  filter(relate == 1) %>% 
  group_by(serialp) %>% 
  filter(n() > 1) %>% 
  censocdev::recode_education(educ_var = educd)

dmf_analysis_sibs <- dmf_analysis_sibs %>% 
  arrange(byear) %>% 
  mutate(birth_order = row_number()) %>% ungroup()
 
model1 <- feols(death_age ~ ownership | byear,  data = dmf_analysis_sibs)
model2 <- feols(death_age ~ ownership + educ_yrs | byear + educ_yrs + urban + occ + race + marst + statefip,  data = dmf_analysis_sibs)
model3 <- feols(death_age ~ ownership + educ_yrs | byear + educ_yrs + urban + occ + race + marst + statefip + serialp + as.factor(birth_order),  data = dmf_analysis_sibs)


results <- tidy(model1) %>% mutate(model = "Baseline") %>% 
  bind_rows(tidy(model2) %>% mutate(model = "Baseline + Controls")) %>% 
  bind_rows(tidy(model3) %>% mutate(model = "Baseline + Controls + Sibling FE")) %>% 
  filter(term == "ownership1")


home_ownership_regression_plot <- results %>% 
    filter(term == "ownership1") %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  mutate(method = "OLS") %>% 
  ggplot(aes(x = model, y = e65, ymin = lower, ymax = upper)) + 
  geom_line() + 
  geom_pointrange(size = 1.1) + 
  theme_cowplot() + 
  ylim(0, 0.9) + 
  labs(x = "Model",
       y = "Additional Years of life") + 
  ggsci::scale_color_lancet() + 
  theme(legend.position = "NULL")

# title = "Relationship between Home Ownership and Longevity",
# subtitle = "CenSoc-DMF cohorts of 1905-1915, Men-Only"


ggsave(plot = home_ownership_regression_plot, "figures/home_ownership_models.pdf", width = 9, height = 5)
```
