==> code/02_homeownership_decade.Rmd <==
---
title: "Estimate homeonership rates by decade"
author: Casey Breen
---

Summary: Estimate home ownership across census decades using IPUMS full count census data (1940, 1930, 1920, and 1900). 

This code produces: 

* Figure 1, Figure 2 

```{r}
## Load helper functions and library all packages used for analysis
library(here)
source(here("code/helpers.R"))
```

## read in 1940 census 

```{r}
## Read in 1940 census 
census_1940_h <- fread("/home/ipums/casey-ipums/IPUMS2019/1940/TSV/H.tsv", select = c("SERIAL", "OWNERSHP")) 
census_1940_p <- fread("/home/ipums/casey-ipums/IPUMS2019/1940/TSV/P.tsv", select = c("SERIALP", "RACE", "AGE", "SEX", "RELATE")) 
census_1940 <- inner_join(census_1940_p, census_1940_h, by = c("SERIALP" = "SERIAL")) %>% janitor::clean_names()

## Recode 1940 census 
census_1940 <- census_1940 %>% 
  janitor::clean_names() %>% 
  mutate(race = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black"
  )) %>% 
  filter(relate == 101 & sex == 1) %>% 
  mutate(age_cat = case_when(
    age %in% 18:24 ~ "[18, 25)",
    age %in% 25:34 ~ "[25, 35)",
    age %in% 35:44 ~ "[35, 45)",
    age %in% 45:54 ~ "[45, 55)",
    age %in% 55:64 ~ "[55, 64)",
    age  >= 65 ~ "65+"
  )) %>% 
  filter(!is.na(age_cat)) %>% 
  filter(!is.na(race))

## Tabulate by race and age 
tabs_1940_bw_age <- census_1940 %>% 
  group_by(age_cat, race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))
  
## Full tabulation 
tabs_1940_bw <- census_1940 %>% 
  group_by(race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))

## Drop big census files  
rm(census_1940_h, census_1940_p)
```

## read in 1930 census 

```{r}
## Read in 1930 Census 
census_1930_h <- fread("/home/ipums/casey-ipums/IPUMS2019/1930/TSV/H.tsv", select = c("SERIAL", "OWNERSHP"))
census_1930_p <- fread("/home/ipums/casey-ipums/IPUMS2019/1930/TSV/P.tsv", select = c("SERIALP", "RACE", "AGE", "SEX", "RELATE"))
census_1930 <- inner_join(census_1930_p, census_1930_h, by = c("SERIALP" = "SERIAL"))


## recode 1940 census 
census_1930 <- census_1930 %>% 
  janitor::clean_names() %>% 
  mutate(race = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black"
  )) %>% 
  filter(relate == 101 & sex == 1) %>% 
  mutate(age_cat = case_when(
    age %in% 18:24 ~ "[18, 25)",
    age %in% 25:34 ~ "[25, 35)",
    age %in% 35:44 ~ "[35, 45)",
    age %in% 45:54 ~ "[45, 55)",
    age %in% 55:64 ~ "[55, 64)",
    age  >= 65 ~ "65+"
  )) %>% 
  filter(!is.na(age_cat)) %>% 
  filter(!is.na(race))

## tabulate by race and age 
tabs_1930_bw_age <- census_1930 %>% 
  group_by(age_cat, race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))
  
## tabulate
tabs_1930_bw <- census_1930 %>% 
  group_by(race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))

## drop big census files  
rm(census_1930_h, census_1930_p)
```

## read in 1920 census 


```{r}
## Read in 1920 Census 
census_1920_h <- fread("/home/ipums/casey-ipums/IPUMS2019/1920/TSV/H.tsv", select = c("SERIAL", "OWNERSHP"))
census_1920_p <- fread("/home/ipums/casey-ipums/IPUMS2019/1920/TSV/P.tsv", select = c("SERIALP", "RACE", "AGE", "SEX", "RELATE"))
census_1920 <- inner_join(census_1920_p, census_1920_h, by = c("SERIALP" = "SERIAL"))


## Recode 1940 census 
census_1920 <- census_1920 %>% 
  janitor::clean_names() %>% 
  mutate(race = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black"
  )) %>% 
  filter(relate == 101 & sex == 1) %>% 
  mutate(age_cat = case_when(
    age %in% 18:24 ~ "[18, 25)",
    age %in% 25:34 ~ "[25, 35)",
    age %in% 35:44 ~ "[35, 45)",
    age %in% 45:54 ~ "[45, 55)",
    age %in% 55:64 ~ "[55, 64)",
    age  >= 65 ~ "65+"
  )) %>% 
  filter(!is.na(age_cat)) %>% 
  filter(!is.na(race))

## Tabulate by race and age 
tabs_1920_bw_age <- census_1920 %>% 
  group_by(age_cat, race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))
  
## Tabulate by race 
tabs_1920_bw <- census_1920 %>% 
  group_by(race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))

## Drop big census files  
rm(census_1920_h, census_1920_p)
```

## read in 1910 census 


```{r}
## Read in 1910 Census 
census_1910_h <- fread("/home/ipums/casey-ipums/IPUMS2019/1910/TSV/H.tsv", select = c("SERIAL", "OWNERSHP"))
census_1910_p <- fread("/home/ipums/casey-ipums/IPUMS2019/1910/TSV/P.tsv", select = c("SERIALP", "RACE", "AGE", "SEX", "RELATE"))
census_1910 <- inner_join(census_1910_p, census_1910_h, by = c("SERIALP" = "SERIAL"))


## Recode 1940 census 
census_1910 <- census_1910 %>% 
  janitor::clean_names() %>% 
  mutate(race = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black"
  )) %>% 
  filter(relate == 101 & sex == 1) %>% 
  mutate(age_cat = case_when(
    age %in% 18:24 ~ "[18, 25)",
    age %in% 25:34 ~ "[25, 35)",
    age %in% 35:44 ~ "[35, 45)",
    age %in% 45:54 ~ "[45, 55)",
    age %in% 55:64 ~ "[55, 64)",
    age  >= 65 ~ "65+"
  )) %>% 
  filter(!is.na(age_cat)) %>% 
  filter(!is.na(race))

## Tabulate by race and age 
tabs_1910_bw_age <- census_1910 %>% 
  group_by(age_cat, race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))
  
## Tabulate by race 
tabs_1910_bw <- census_1910 %>% 
  group_by(race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))

## Drop big census files  
rm(census_1910_h, census_1910_p)
```

## read in 1900 census 

```{r}
## Read in 1900 Census 
census_1900_h <- fread("/home/ipums/casey-ipums/IPUMS2019/1900/TSV/H.tsv", select = c("SERIAL", "OWNERSHP"))
census_1900_p <- fread("/home/ipums/casey-ipums/IPUMS2019/1900/TSV/P.tsv", select = c("SERIALP", "RACE", "AGE", "SEX", "RELATE"))
census_1900 <- inner_join(census_1900_p, census_1900_h, by = c("SERIALP" = "SERIAL"))


## Recode 1940 census 
census_1900 <- census_1900 %>% 
  janitor::clean_names() %>% 
  mutate(race = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black"
  )) %>% 
  filter(relate == 101 & sex == 1) %>% 
  mutate(age_cat = case_when(
    age %in% 18:24 ~ "[18, 25)",
    age %in% 25:34 ~ "[25, 35)",
    age %in% 35:44 ~ "[35, 45)",
    age %in% 45:54 ~ "[45, 55)",
    age %in% 55:64 ~ "[55, 64)",
    age  >= 65 ~ "65+"
  )) %>% 
  filter(!is.na(age_cat)) %>% 
  filter(!is.na(race))

## Tabulate by race and age 
tabs_1900_bw_age <- census_1900 %>% 
  group_by(age_cat, race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))
 
## Tabulate by race
tabs_1900_bw <- census_1900 %>% 
  group_by(race) %>% 
  summarize(ownershp_avg = mean(ownershp == 10))

## Drop big census files  
rm(census_1900_h, census_1900_p)
```

## Create homeownership plot by race 

```{r}
## Combine estimates  
bw_tabs <- tabs_1940_bw %>% mutate(year = 1940) %>% 
  bind_rows(tabs_1930_bw %>% mutate(year = 1930)) %>%
  bind_rows(tabs_1920_bw %>% mutate(year = 1920)) %>% 
  bind_rows(tabs_1910_bw %>% mutate(year = 1910)) %>% 
  bind_rows(tabs_1900_bw %>% mutate(year = 1900))
  
## Add on additional estimates from Collins and Margo (2011)
bw_external_source <- tribble(
  ~race, ~ownershp_avg, ~year,
  "White",   .512, 1890, 
  "Black",   .193, 1890, 
  "White",   .567, 1880,
  "Black",   .137, 1880, 
  "White",   .565, 1870, 
  "Black",   .077, 1870
)

## create homeownership plot by age 
bw_homeownership_plot <- bw_tabs %>%
  bind_rows(bw_external_source) %>% 
  mutate(label = if_else(year == max(year), as.character(race), NA_character_)) %>%
  mutate(race = case_when(
    race == "Black" ~ "Black Americans",
    race == "White" ~ "White Americans"
  )) %>% 
  ggplot(aes(x = year, y = ownershp_avg*100, linetype = race, shape = race)) + 
  geom_line(size = 1) + 
  geom_point(size = 3.5, fill = "white") + 
  scale_y_continuous(breaks = scales::pretty_breaks(n = 10), limits = c(0, 60)) + 
  scale_x_continuous(breaks = scales::pretty_breaks(n = 8), limits = c(1870, 1945)) + 
  cowplot::theme_cowplot() + 
  scale_shape_manual(values = c(21, 23)) + 
  theme(legend.position = "NULL",
        legend.title=element_blank()) + 
  geom_text_repel(aes(label = label),
                  nudge_x = 5,
                  na.rm = TRUE,
                segment.color = NA) + 
  labs(x = "Year",
       y = "Homeownership Rate (%)") + 
  theme(axis.title.y = element_text(face="bold", colour = "black"),
        axis.title.x = element_text(face="bold", colour = "black")) 


## save plot 
ggsave(plot = bw_homeownership_plot, filename = here("figures/fig1_home_ownership_by_decade.png"), width = 8, height = 5)
ggsave(plot = bw_homeownership_plot, filename = here("figures/fig1_home_ownership_by_decade.pdf"), width = 8, height = 5)
```


## Create homeownership plot by race X age 

```{r}
## Combine estimates 
bw_tabs_age <- tabs_1940_bw_age %>% mutate(year = 1940) %>% 
  bind_rows(tabs_1930_bw_age %>% mutate(year = 1930)) %>%
  bind_rows(tabs_1920_bw_age %>% mutate(year = 1920)) %>% 
  bind_rows(tabs_1910_bw_age %>% mutate(year = 1910)) %>% 
  bind_rows(tabs_1900_bw_age %>% mutate(year = 1900))

## Create labels 
bw_tabs_age <- bw_tabs_age %>% 
  mutate(race = case_when(
    race == "Black" ~ "Black",
    race == "White" ~ "White"
  )) %>% 
  mutate(age_cat = case_when(
    age_cat == "65+" ~ "65+     ",
    TRUE ~ age_cat
  ))

## Create homeownership rates plot 
# Separate data for Black race
home_ownership_plot_black <- bw_tabs_age %>%
  filter(race == "Black") %>%
  mutate(label = if_else(year == max(year), as.character(age_cat), NA_character_)) %>%
  ggplot(aes(x = year, y = ownershp_avg * 100, color = age_cat, shape = race)) + 
  geom_line(size = 1) + 
  geom_point(size = 3, fill = "white") + 
  cowplot::theme_cowplot() + 
  scale_color_brewer(palette = "Dark2") +
  theme(legend.position = "none") + 
  scale_shape_manual(values = c(21, 23)) + 
  scale_y_continuous(breaks = scales::pretty_breaks(n = 20), lim = c(0, 76)) + 
  labs(x = "Year", y = "Homeownership Rate (%)", title = "Black") + 
  geom_text_repel(aes(label = label), nudge_x = 7, na.rm = TRUE, 
                segment.color = NA) + 
  xlim(1900, 1947) + 
  theme(strip.background = element_blank()) + 
  panel_border() + 
  theme(axis.title.y = element_text(face = "bold", colour = "black"),
        axis.title.x = element_text(face = "bold", colour = "black"))

# Separate data for White race
home_ownership_plot_white <- bw_tabs_age %>%
  filter(race == "White") %>%
  mutate(label = if_else(year == max(year), as.character(age_cat), NA_character_)) %>%
  ggplot(aes(x = year, y = ownershp_avg * 100, color = age_cat, shape = race)) + 
  geom_line(size = 1) + 
  geom_point(size = 3, fill = "white") + 
  cowplot::theme_cowplot() + 
  scale_color_brewer(palette = "Dark2") +
  theme(legend.position = "none") + 
  scale_shape_manual(values = c(21, 23)) + 
  scale_y_continuous(breaks = scales::pretty_breaks(n = 20), lim = c(0, 76)) + 
  labs(x = "Year", y = "", title = "White") + 
  geom_text_repel(aes(label = label), nudge_x = 7, na.rm = TRUE, segment.color = NA) + 
  xlim(1900, 1947) + 
  theme(strip.background = element_blank()) + 
  panel_border() + 
  theme(axis.title.y = element_text(face = "bold", colour = "black"),
        axis.title.x = element_text(face = "bold", colour = "black"))

# Combine the two plots using plot_grid
home_ownership_rates_plot <- cowplot::plot_grid(home_ownership_plot_black, home_ownership_plot_white, ncol = 2)

## Save plot 
ggsave(plot = home_ownership_rates_plot, filename = here("figures/fig2_home_ownership_by_decade_age.png"), width = 9, height = 5)
ggsave(plot = home_ownership_rates_plot, filename = here("figures/fig2_home_ownership_by_decade_age.pdf"), width = 9, height = 5)
```
