==> code/05_pooled_analysis.Rmd <==
---
title: "Association between homeownership and longevity"
author: Casey Breen
---

Summary: Fit and visualize regression models on the association of homeownership on longevity. 

This code generates: 

- Figure 6
- Table 2

```{r}
## Load helper functions and library all packages used for analysis
library(here)
source(here("code/helpers.R"))
```

## read in data 

```{r}
## read in dmf
dmf <- fread("/censoc/data/censoc_linked_with_census/1940_v2.1/censoc_dmf_v2_1_linked_with_census.csv")
```

## Association between home ownership and longevity 

```{r}
## dmf household head
dmf_hhead <- dmf %>%
  filter(relate == 101) %>%
  mutate(ownership = relevel(as.factor(ownershp), ref = "20")) %>%
  filter(ownership != 0) %>%
  filter(bpl < 15000) %>%
  filter(byear %in% 1905:1915 & age %in% 22:37)


dmf_hhead <- dmf_hhead %>% 
  mutate(ownership = relevel(as.factor(ownershp), ref = "20")) %>% 
  mutate(racec = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black",
    TRUE ~ "Other"
  )) %>% 
  mutate(educ = as.numeric(educ)) %>% 
  recode_education(educ_var = "educ") %>% 
  filter(!is.na(educ_yrs))

## pooled effects white 
pooled_effect_model_white <- fixest::feols(death_age ~ ownership | byear, dmf_hhead %>% filter(race == 100))

pooled_effect_df_white <- pooled_effect_model_white %>%
  tidy() %>% 
  mutate(
    type = "White",
    model = "No Controls"
  )

## pooled effects whites with controls
pooled_effect_controls_model_white <- feols(death_age ~ ownership + educ_yrs  + urban | byear  + occ  + marst + statefip, dmf_hhead %>% filter(race == 100))

pooled_effect_controls_df_white <- pooled_effect_controls_model_white %>%
  tidy() %>% 
  mutate(
    type = "White",
    model = "Controls"
  )

## pooled effects blacks
pooled_effect_model_black <- fixest::feols(death_age ~ ownership | byear, dmf_hhead %>% filter(race == 200))

pooled_effect_black_df <- pooled_effect_model_black %>% 
  tidy() %>% 
  mutate(
    type = "Black",
    model = "No Controls"
  )

## pooled effects blacks with controls
pooled_effect_controls_model_black <- feols(death_age ~ ownership + educ_yrs + urban | byear  + occ  + marst + statefip, dmf_hhead %>% filter(race == 200))

pooled_effect_controls_black_df <- pooled_effect_controls_model_black %>% 
  tidy() %>% 
  mutate(
    type = "Black",
    model = "Controls"
  )
```

## Visualize regression coefficients

```{r}
## pooled effects
bw_differences_homeown_longevity <- pooled_effect_df_white %>%
  bind_rows(pooled_effect_controls_df_white) %>%
  bind_rows(pooled_effect_black_df) %>%
  bind_rows(pooled_effect_controls_black_df) %>%
  filter(term == "ownership10") %>% 
  ggplot(aes(
    x = reorder(model, -estimate), y = estimate, color = rev(model), shape = type,
    ymin = estimate + 1.96 * std.error,
    ymax = estimate - 1.96 * std.error
  )) +
  geom_pointrange(size = 0.7,  fill = "white", position = position_dodge2(.16)) +
  theme_cowplot() +
  scale_y_continuous(labels = function(x) {
      ifelse(x < 0, gsub("-", "\u2013", format(x)), format(x))
    }, lim= c(-0.2, 1)) + 
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") +
  annotate("text", x = 1.5, y = -.05, label = "Renters", fontface = "italic") +
  scale_shape_manual(values = c(21, 23)) + 
  scale_color_manual(values = cud) + 
  labs(
    x = "",
    y = "Difference in Life Expectancy \n (years)",
    title = ""
  ) +
  theme(
    legend.position = "bottom",
    legend.title = element_blank()
  ) +
  guides(linetype = "none", color = "none") + 
  theme(axis.title.y = element_text(face="bold", colour = "black"),
        axis.text.x = element_text(face="bold", colour = "black")) 

## save plot
ggsave(bw_differences_homeown_longevity, filename = here("figures/fig5_homeownership_association_bw.png"), width = 6, height = 4)
ggsave(bw_differences_homeown_longevity, filename = here("figures/fig5_homeownership_association_bw.pdf"), width = 6, height = 4)
```

## Create regression table 

```{r}
## fit full regression models 

etable(pooled_effect_model_white, pooled_effect_controls_model_white, pooled_effect_model_black, pooled_effect_controls_model_black, tex = T, file = here("tables/tableS2_regtable_homeownership_bw.tex"), replace = T)
```
