==> code/06_sibling_analysis.Rmd <==
---
title: "Home Ownership Sibling Analysis"
author: Casey Breen
---

Summary: Create causal estimates of association between home ownership and longevity using sibling-based fixed effects models. 

This code generates: 

- Figure 6, 7, 9 
- Table S2, S3, S4

## Library packages 

```{r}
## Load helper functions and library all packages used for analysis
library(here)
source(here("code/helpers.R"))

## read in sibling mortality data 
dmf_sibs <- read_rds(here("data/dmf_sibs.rds")) 
```

## Figure 7 — Standard models (pooled, pooled adjusted, sib-FE adjusted)

```{r}
## fit full regression models 
model1 <- feols(death_age ~ ownership | byear,  data = dmf_sibs)
model2 <- feols(death_age ~ ownership + educ_yrs + racec + urban | byear  + occ  + marst + statefip, data = dmf_sibs)
model3 <- feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = dmf_sibs)
model4 <- feols(death_age ~ ownership  | byear + serialp_1920 + birth_order, data = dmf_sibs)

## model 1 
results <- tidy(model1) %>% 
  mutate(model = "No Controls") %>% 
  add_row(model = "Controls", estimate = NA, std.error = NA) %>% 
  add_row(model = "Controls + Sibling FE", estimate = NA, std.error = NA) %>% 
  mutate(model = factor(model, levels= c("No Controls", "Controls", "Controls + Sibling FE")))

## model 1 results 
home_ownership_regression_plot1 <- results %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  mutate(method = "OLS") %>% 
  ggplot(aes(x = model, y = e65, ymin = lower, ymax = upper, color = model)) + 
  geom_pointrange(size = 1,  fill = "white", shape = 22) + 
  theme_cowplot() + 
  ylim(-0.2, 0.9) + 
  labs(x = "",
       y = "Additional Years of Life") + 
  scale_color_manual(values = cud) + 
  theme(legend.position = "NULL") +
   annotate("text", x = 2.3, y = -.05, label = "Renters", fontface = 'italic') + 
  geom_hline(yintercept = 0, color = "grey", linetype =  "dashed")

## fit model 2 
results <- tidy(model1) %>% mutate(model = "No Controls") %>% 
  bind_rows(tidy(model2) %>% mutate(model = "Controls")) %>% 
 add_row(model = "Controls + Sibling FE", estimate = NA, std.error = NA) %>% 
  mutate(model = factor(model, levels= c("No Controls", "Controls", "Controls + Sibling FE")))

## visualize model 2 
home_ownership_regression_plot2 <- results %>% 
  filter(term == "ownership10" | is.na(term)) %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  mutate(method = "OLS") %>% 
  ggplot(aes(x = model, y = e65, ymin = lower, ymax = upper, color = model)) + 
  geom_pointrange(size = 1,  fill = "white", shape = 22) + 
  theme_cowplot() + 
  ylim(-0.2, 0.9) + 
  labs(x = "",
       y = "Additional Years of Life") + 
  scale_color_manual(values = cud) + 
  theme(legend.position = "NULL") +
   annotate("text", x = 2.3, y = -.05, label = "Renters", fontface = 'italic') + 
  geom_hline(yintercept = 0, color = "grey", linetype =  "dashed")

## fit model 3 
results <- tidy(model1) %>% mutate(model = "No Controls") %>% 
  bind_rows(tidy(model2) %>% mutate(model = "Controls")) %>% 
  bind_rows(tidy(model3) %>% mutate(model = "Controls + Sibling FE")) %>% 
  mutate(model = factor(model, levels= c("No Controls", "Controls", "Controls + Sibling FE")))

## visualize model 3 
home_ownership_regression_plot3 <- results %>% 
    filter(term == "ownership10") %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  mutate(method = "OLS") %>% 
  ggplot(aes(x = model, y = e65, ymin = lower, ymax = upper, color = model)) + 
  geom_pointrange(size = 1,  fill = "white", shape = 22) + 
  theme_cowplot() + 
  labs(x = "",
       y = "Additional Years of Life") + 
  scale_color_manual(values = cud) + 
  theme(legend.position = "NULL") +
  annotate("text", x = 2.3, y = -.05, label = "Renters", fontface = 'italic') + 
  geom_hline(yintercept = 0, color = "grey", linetype =  "dashed") + 
  theme(axis.text.x = element_text(face="bold", colour = "black"),
        axis.title.y = element_text(face="bold", colour = "black")) + 
    scale_y_continuous(labels = function(x) {
      ifelse(x < 0, gsub("-", "\u2013", format(x)), format(x))
    }, lim= c(-0.2, 0.9))

## save plots 
ggsave(plot = home_ownership_regression_plot1, here("figures/reserve_figures/home_ownership_sib_models1.png"), width = 6, height = 4)
ggsave(plot = home_ownership_regression_plot2, here("figures/reserve_figures/home_ownership_sib_models2.png"), width = 6, height = 4)
ggsave(plot = home_ownership_regression_plot3, here("figures/fig6_home_ownership_sib_models3.png"), width = 6, height = 4)
ggsave(plot = home_ownership_regression_plot3, here("figures/fig6_home_ownership_sib_models3.pdf"), width = 6, height = 4)
```

## Table 2 — Full Regression Models 

```{r}
etable(model1, model2, model4, model3, tex = T, file = here("tables/tableS3_regtable_sibling_models.tex"), replace = T)
```


## Causal ABE models — Conservative Subsample (pooled, pooled adjusted, sib-FE adjusted)

```{r}
## Restrict to conservative subsample for both linkages 
dmf_sibs_conservative <- dmf_sibs %>% 
  filter(link_abe_exact_conservative_1940_dmf == 1 & link_abe_exact_conservative_1920_1940 == 1) %>% 
  group_by(serialp_1920) %>% 
  filter(n() > 1) %>% 
  arrange(byear) %>% 
  mutate(birth_order = row_number()) %>%
  ungroup()
 
## fit full regression models 
model1 <- feols(death_age ~ ownership | byear,  data = dmf_sibs_conservative)
model2 <- feols(death_age ~ ownership + educ_yrs + racec + urban | byear  + occ  + marst + statefip, data = dmf_sibs_conservative)
model3 <- feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = dmf_sibs_conservative)
model4 <- feols(death_age ~ ownership  | byear + serialp_1920 + birth_order, data = dmf_sibs_conservative)


## combine all results into one data.frame
results <- tidy(model1) %>% mutate(model = "No Controls") %>% 
  bind_rows(tidy(model2) %>% mutate(model = "Controls")) %>% 
  bind_rows(tidy(model3) %>% mutate(model = "Controls + Sibling FE")) %>% 
  mutate(model = factor(model, levels= c("No Controls", "Controls", "Controls + Sibling FE")))

## visualize model 3 
home_ownership_regression_conservative_plot <- results %>% 
    filter(term == "ownership10") %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  mutate(method = "OLS") %>% 
  ggplot(aes(x = model, y = e65, ymin = lower, ymax = upper, color = model)) + 
  geom_pointrange(size = 1,  fill = "white", shape = 22) + 
  theme_cowplot() + 
  ylim(-0.2, 0.9) + 
  labs(x = "Model",
       y = "Additional Years of Life") + 
  scale_color_manual(values = cud) + 
  theme(legend.position = "NULL") +
  annotate("text", x = 2.3, y = -.05, label = "Renters", fontface = 'italic') + 
  geom_hline(yintercept = 0, color = "grey", linetype =  "dashed")

## Save plot 
ggsave(plot = home_ownership_regression_conservative_plot, here("figures/fig9_home_ownership_sib_models_conservative.png"), width = 6, height = 4)
```

## discordant siblings 

```{r}
## Number of discordant sibships 
dmf_sibs %>% 
  group_by(serialp_1920) %>% 
  summarize(distinct = n_distinct(ownershp)) %>% 
  count(distinct) %>% 
  summarize(prop = n / sum(n))
```

## Sensitivity analysis 

Create different, more restrictive subsamples and rerun model 3 on each subgroup. The idea is by making such restrictions (e.g., sibling living in the same county) we are able to control for potential confounding. 

```{r}
## byear restriction 
byear_restrict_df <- dmf_sibs %>%
   group_by(serialp_1920) %>% 
   mutate(byear_range = max(byear) - min(byear)) %>% 
   filter(byear_range < 6)

model_similar_byear <- tidy(feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = byear_restrict_df)) %>% mutate(model = "Birth Year Restriction")

## county restriction 
county_restriction_df <- dmf_sibs %>% 
  group_by(serialp_1920) %>% 
  filter(n_distinct(countyicp) == 1) %>% 
  ungroup()

model_county <- tidy(feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = county_restriction_df)) %>% mutate(model = "County Restriction")

## marital status restriction 
marst_restriction_df <- dmf_sibs %>% 
  group_by(serialp_1920) %>% 
  filter(n_distinct(marst) == 1) %>% 
  ungroup()

marst_similar_model <- tidy(feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = marst_restriction_df)) %>% mutate(model = "Marital Status Restriction")


## education status restriction 
educ_restriction_df <- dmf_sibs %>% 
  group_by(serialp_1920) %>% 
  mutate(educ_range = max(educ_yrs) - min(educ_yrs)) %>% 
  filter(educ_range < 3) %>% 
  ungroup()

educ_similar_model <- tidy(feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = educ_restriction_df)) %>% mutate(model = "Education Restriction")

## full restrictions 
full_restrictions <- dmf_sibs %>% 
  group_by(serialp_1920) %>% 
  mutate(byear_range = max(byear) - min(byear)) %>% 
  mutate(educ_range = max(educ_yrs) - min(educ_yrs)) %>% 
  filter(byear_range < 6) %>% 
  filter(educ_range < 3) %>% 
  filter(n_distinct(countyicp) == 1) %>% 
  filter(n_distinct(marst) == 1)

full_restrictions_model <- tidy(feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = full_restrictions)) %>% mutate(model = "All Restrictions ")

## combine estimates into one data frame
results_robustness <- model_similar_byear %>% 
  bind_rows(model_county) %>% 
  bind_rows(marst_similar_model) %>% 
  bind_rows(educ_similar_model) %>% 
  bind_rows(full_restrictions_model)

## sensitivity analysis 
sensitivity_analysis_plots <- results_robustness %>% 
      filter(term == "ownership10") %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  mutate(method = "OLS") %>% 
  ggplot(aes(y = model, x = e65, xmin = lower, xmax = upper)) + 
  geom_line() + 
  geom_pointrange(size = 0.7, shape = 22, fill = "white") + 
  theme_cowplot() + 
  xlim(-.3, 0.9) + 
  labs(x = "Additional Years of Life",
       y = "") + 
  ggsci::scale_color_lancet() + 
  geom_vline( xintercept = 0.31, linetype = "dashed") + 
  theme(legend.position = "NULL") + 
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  annotate("text", x = .03, y = 2.5, label = "Renters", fontface = 'italic', angle='90') + 
  geom_vline(xintercept = 0, color = "grey", linetype =  "dashed") + 
  annotate("text", x = .35, y = 2.5, label = "Full Sample", fontface = 'italic', angle='90') + 
    theme(axis.text.y = element_text(face="bold", colour = "black"),
        axis.title.x = element_text(face="bold", colour = "black")) 

## save plot 
ggsave(plot = sensitivity_analysis_plots, here("figures/fig7_home_ownership_sensitivity_analyses.png"), width = 7.7, height = 6)
ggsave(plot = sensitivity_analysis_plots, here("figures/fig7_home_ownership_sensitivity_analyses.pdf"), width = 7.7, height = 6)

```
