==> code/07_homeownership_homevalue.Rmd <==
---
title: "Home Ownership and Mortality by Quintile"
author: Casey Breen
---

Summary: identify how the quantile of wealth in 1940 

## Library packages 

```{r}
## Load helper functions and library all packages used for analysis
library(here)
source(here("code/helpers.R"))

## Read in sibling file 
dmf_analysis_sibs <- read_rds(here("data/dmf_sibs.rds")) %>% 
  janitor::clean_names()
```

## Calculate homeownership value quartile

```{r}
## find values of households 
hh <- fread("/ipums-repo2019/1940/TSV/H.tsv")

## restrict to homes with non-missing values for ownership 
hh_homes <- hh %>% 
  filter(OWNERSHP == 10)
 
## filter to only homes with a non-missing values 
hh_homes <- hh_homes %>% 
  filter(VALUEH != 9999998)

## split into quartiles 
hh_homes <- hh_homes %>% 
  group_by(STATEFIP) %>% 
  mutate(hvalue_state = ntile(VALUEH, 4)) %>% 
  ungroup()
```

## Link analysis file onto CenSoc file 

```{r}
## dmf sibs 
dmf_analysis_sibs <- dmf_analysis_sibs %>% 
  left_join(hh_homes %>% select(serialp_1940 = SERIAL, hvalue_state))

## recode dmf sibs 
dmf_analysis_sibs <- dmf_analysis_sibs %>% 
  ungroup() %>% 
    mutate(ownership = relevel(as.factor(ownershp), ref = "20")) %>% 
  filter(ownership != 10 | valueh != 9999998) %>% 
  mutate(ownership_value = case_when(
    ownership == 10 & hvalue_state == 1 ~ 1,
    ownership == 10 & hvalue_state == 2 ~ 2,
    ownership == 10 & hvalue_state == 3 ~ 3,
    ownership == 10 & hvalue_state == 4 ~ 4,
    TRUE ~ 0
    )) %>% 
   mutate(racec = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black",
    TRUE ~ "Other"
  )) %>% 
  group_by(serialp_1920) %>% 
  arrange(byear) %>% 
  mutate(birth_order = row_number()) %>% 
  ungroup() %>% 
  mutate(racec = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black",
    TRUE ~ "Other"
  ))

## dmf sibs 
dmf_analysis_sibs <- dmf_analysis_sibs %>% 
  group_by(serialp_1920) %>% 
  mutate(n = n()) %>% 
  filter(n > 1) %>% 
  ungroup()

## run models 
model1 <- tidy(feols(death_age ~ as.factor(ownership_value)  | byear, data = dmf_analysis_sibs)) %>% mutate(model = "No controls")
model2 <- tidy(feols(death_age ~ as.factor(ownership_value)  + educ_yrs + racec + urban | byear  + occ  + marst + statefip, data = dmf_analysis_sibs)) %>% mutate(model = "Controls")
model3 <- tidy(feols(death_age ~ as.factor(ownership_value) + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = dmf_analysis_sibs)) %>% mutate(model = "Controls + sibling FE")

## effect heterogeneity 
home_ownership_homevalue_plot_1models <- model3 %>% 
  filter(str_detect(pattern = "as.factor", string = term)) %>% 
  mutate(term = case_when(
    term == "as.factor(ownership_value)1" ~ "Q1 (Low)",
    term == "as.factor(ownership_value)2" ~ "Q2",
    term == "as.factor(ownership_value)3" ~ "Q3",
    term == "as.factor(ownership_value)4" ~ "Q4 (High)"
  )) %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  ggplot(aes(x = term, y = e65, ymin = lower, ymax = upper, group = model, color = model, shape = model)) + 
    geom_line(alpha = 0.6, linetype = "dashed") + 
  geom_pointrange(size = 0.7,  fill = "white", shape = 22, position = position_dodge2(.1)) + 
  theme_cowplot() + 
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") + 
  labs(x = "Home Value Quartile",
       y = "Additional Years of Life") + 
  #  ggsci::scale_color_lancet() + 
  scale_color_manual(values = c("Controls + Sibling FE" = "#008B45FF")) + 
  # scale_color_viridis_d(option = "m") + 
  theme(legend.position = "bottom") + 
  annotate("text", x = 3.5, y = -.05, label = "Renters",
           parse = TRUE)

## effect heterogeneity 
home_ownership_homevalue_plot_3models <- model1 %>% 
  bind_rows(model2) %>% 
  bind_rows(model3) %>% 
    filter(str_detect(pattern = "as.factor", string = term)) %>% 
  mutate(term = case_when(
    term == "as.factor(ownership_value)1" ~ "Q1 (low)",
    term == "as.factor(ownership_value)2" ~ "Q2",
    term == "as.factor(ownership_value)3" ~ "Q3",
    term == "as.factor(ownership_value)4" ~ "Q4 (high)"
  )) %>% 
  mutate(model = factor(model, levels= c("No controls", "Controls", "Controls + sibling FE"))) %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  ggplot(aes(x = term, y = e65, ymin = lower, ymax = upper, group = model, color = model, shape = model)) + 
    geom_line(alpha = 0.6, linetype = "dashed") + 
  geom_pointrange(size = 0.7,  fill = "white", shape = 22, position = position_dodge2(.1)) + 
  theme_cowplot() + 
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") + 
  scale_color_manual(values = cud) + 
  labs(x = "Home Value Quartile",
       y = "Additional Years of Life") + 
  theme(legend.position = "bottom") + 
  annotate("text", x = 3.5, y = -.05, label = "Renters", fontface = "italic") + 
  theme(axis.title.y = element_text(face="bold", colour = "black"),
  axis.title.x = element_text(face="bold", colour = "black"),
  legend.title=element_blank()) 

## homeownership plot 1 
ggsave(plot = home_ownership_homevalue_plot_3models, here("figures/fig8_home_ownership_homevalue_plot_3models.png"), width = 6, height = 4)
ggsave(plot = home_ownership_homevalue_plot_3models, here("figures/fig8_home_ownership_homevalue_plot_3models.pdf"), width = 6, height = 4)

## homeownership plot 3 
ggsave(plot = home_ownership_homevalue_plot_1models, here("figures/reserve_figures/home_ownership_homevalue_plot_1models.png"), width = 6, height = 4)
```


## create regression table 

```{r}
## rerun models 
model1 <- feols(death_age ~ as.factor(ownership_value)  | byear, data = dmf_analysis_sibs)
model2 <- feols(death_age ~ as.factor(ownership_value)  + educ_yrs + racec + urban | byear  + occ  + marst + statefip, data = dmf_analysis_sibs)
model3 <- feols(death_age ~ as.factor(ownership_value) + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = dmf_analysis_sibs)

etable(model1, model2, model3, tex = T, file = here("tables/tableS4_regtable_homeownership_quartile.tex"), replace = T)
```
