==> code/08_gompertztrunc_analysis.Rmd <==
---
title: "Gompertztrunc Analyses"
author: Casey Breen 
---


```{r}
## Load helper functions and library all packages used for analysis
library(here)
source(here("code/helpers.R"))

## read in sibling mortality data 
dmf_sibs <- read_rds(here("data/dmf_sibs.rds"))

## recode siblings 
dmf_sibs <- dmf_sibs %>% 
  mutate(ownership_dummy = case_when(
    ownership == 10 ~ 1,
    TRUE ~ 0
  ))
```


```{r}
## run gompertz model 
gompertz_model <- gompertztrunc::gompertz_mle(data = dmf_sibs, death_age ~ ownership_dummy)

## convert gompertz model to hazards
gompertztrunc::convert_hazards_to_ex(gompertz_model$results, use_model_estimates = T)
```

```{r}
## linear model 
gompertz_model_lm <- lm(data = dmf_sibs, death_age ~ ownership_dummy + as.factor(byear))

## 17% increase 
0.7416317/0.6366
```
