==> code/10_construct_weights.Rmd <==
---
title: "Construct survey weights"
author: Casey Breen
---

Summary: Construct survey weights and rerun main regressions 

```{r}
## Load helper functions and library all packages used for analysis
library(here)
source(here("code/helpers.R"))

## dmf sibs 
dmf_sibs <- read_rds(here("data/dmf_sibs.rds")) 

## read in marginal totals 
marginals <- read_csv(here("data/marginals.csv")) 

## read data dictionary initiative (DDI) file with varnames and var labels
ipums_ddi <- ipumsr::read_ipums_ddi(here("data/fullcount.ddi.xml"))
```


```{r}
## recode variables 
dmf_sibs <- dmf_sibs %>% 
  janitor::clean_names(case = "all_caps") %>%
  ipumsr::ipums_collect(ipums_ddi, var_attrs = c("val_labels", "var_label", "var_desc")) %>%
  janitor::clean_names()  %>% 
  mutate(
  hs = case_when(
    educ >= 60 & educ < 998 ~ 1,
    TRUE ~ 0
  ), 
  rural = case_when(
    urban == 1 ~ 1,
    TRUE ~ 0
  ),
  black = case_when( 
    race == 200 ~ 1,
    TRUE ~ 0
  ), white = case_when(
    race == 100 ~ 1,
    TRUE ~ 0
  ),
  homeown = case_when(
    ownershp == 10 ~ 1, 
    TRUE ~ 0
  ),
  p_hh_head = case_when(
    relate == 101 ~ 1, 
    TRUE ~ 0
  ),
  p_hh_head = case_when(
    relate == 101 ~ 1, 
    TRUE ~ 0
  ),
  educ_level = case_when(
    educ <= 50 ~ "< High School",
    educ %in% 60:90 ~ "High School or some college",
    educ %in% 100 ~ "Bachelors Degree",
    educ %in% 110:116 ~ "Advanced Degree",
    TRUE ~ "Other"
  ), 
  sei_recode = case_when(
    sei %in% 1:9 ~ "sei_1_9",
    sei %in% 10:14 ~ "sei_10_14",
    sei %in% 15:25 ~ "sei_15_25",
    sei >= 26 ~      "sei_26+",
    TRUE ~ "Other",
  ),
  marital_status = case_when(
      marst %in% 1:2 ~ "married",
      TRUE ~ "not married" ),
  race_recode = case_when(
    race == 100 ~ "White",
    race == 200 ~ "Black",
    TRUE ~ "Other"
  )) %>% 
  mutate(rural = case_when(
    rural == 1 ~ "Rural",
    TRUE ~ "Urban"
  ),
  homeown = case_when(
    homeown == 1 ~ "Home Owner",
    TRUE ~ "Not Home Owner"
  ),
  region_string = as_factor(region)) 
```


```{r}
## get around 
marginals_round <- marginals %>% 
  mutate(proportion = case_when(
    variable != "region_string" ~ round(proportion, 3),
    TRUE ~ round(proportion, 4))) %>% 
  filter(variable %in% c("region_string", "homeown", "rural", "race_recode", "sei_recode", "marital_status", "educ_level"))

## create weights 
dmf_sibs <- autumn::harvest(data = dmf_sibs %>% mutate(region_string = as.character(region_string)), target = marginals_round) 
```

```{r}
# Define regression models
model1_unweighted = feols(death_age ~ ownership | byear, data = dmf_sibs)
model1_weighted = feols(death_age ~ ownership | byear, weights = dmf_sibs$weights, data = dmf_sibs)
  
model2_unweighted = feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip, data = dmf_sibs)
model2_weighted = feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip, weights = dmf_sibs$weights, data = dmf_sibs)
  
model3_unweighted = feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, data = dmf_sibs)
model3_weighted = feols(death_age ~ ownership + educ_yrs + racec + urban | byear + occ + marst + statefip + serialp_1920 + birth_order, weights = dmf_sibs$weights, data = dmf_sibs)


## fit model 3 
results <- tidy(model1_unweighted) %>% mutate(model = "No Controls", weighted = "unweighted") %>% 
  bind_rows(tidy(model2_unweighted) %>% mutate(model = "Controls", weighted = "unweighted")) %>% 
  bind_rows(tidy(model3_unweighted) %>% mutate(model = "Controls + Sibling FE", weighted = "unweighted")) %>% 
  bind_rows(tidy(model1_weighted) %>% mutate(model = "No Controls", weighted = "weighted")) %>% 
  bind_rows(tidy(model2_weighted) %>% mutate(model = "Controls", weighted = "weighted")) %>% 
  bind_rows(tidy(model3_weighted) %>% mutate(model = "Controls + Sibling FE", weighted = "weighted")) %>% 
  mutate(model = factor(model, levels= c("No Controls", "Controls", "Controls + Sibling FE")))


# Create the final plot with a legend
homeownership_weight_plots <- results %>% 
    filter(term == "ownership10") %>% 
  mutate(e65 = estimate,
         lower = estimate - 1.96*std.error,
         upper = estimate + 1.96*std.error) %>% 
  mutate(method = "OLS") %>% 
  ggplot(aes(x = model, y = e65, ymin = lower, ymax = upper, color = model,  shape = weighted, linetype = weighted)) + 
  geom_pointrange(size = 1,  fill = "white", position = position_dodge(width = 0.3)) +  # Separate estimates
  theme_cowplot() + 
  ylim(-0.2, 0.9) + 
  labs(x = "Model",
       y = "Additional Years of Life") + 
  scale_color_manual(values = cud) + 
  scale_shape_manual(values = c(22, 24)) +  # Shapes 20 and 21
  theme(legend.position = "bottom", legend.title=element_blank()) +  # Place the legend at the bottom
  annotate("text", x = 2.3, y = -.05, label = "Renters", fontface = 'italic') + 
  geom_hline(yintercept = 0, color = "grey", linetype =  "dashed") + 
  guides(colour = "none")



# Save the final plot
ggsave(plot = homeownership_weight_plots, here("figures/fig10_weighted_homeownership_plot.png"), width = 6, height = 4)

```
